[feat] Entity 속성 수정
- Long type의 id를 Integer type으로 변경
  - PostgreSQL에서 long 타입을 지원하지 않음
- 복합키 키 순서 변경

+++ src/main/java/study/practice/domain/model/ReservableRoomId.java
@@ -16,7 +16,7 @@ import java.util.Objects;
 public class ReservableRoomId implements Serializable {
 
     @Column(name = "room_id")
-    private Long roomId; //@MapsId("room_id")로 매핑
+    private Integer roomId; //@MapsId("room_id")로 매핑
 
     @Column(name = "reserved_date")
     private LocalDate reservedDate;
